Solve serialization problem for Jackson ObjectMapper
Code as is doesn't work on a distributed system due to serialization of the ObjectMapper. There are 2 solutions for this. Abstract the ObjectMapper within a singleton Object class on the driver. This has performance implications that everything going through that has to be on the driver node. This may or may not be an issue depending on where you're at in processing. If this is the very first step, then perhaps the driver is the best place for this. While good in that case, it still keeps receiving and processing concerns coupled and a repartition of the stream won't help add processing capacity if this is the bottleneck.

The alternative approach, as demonstrated by the code, is to have each execute create the object using mapPartitions to alleviate heavy object creation, i.e., so we don't create the object for each record and instead, once per partition.